okay, so the computation survives through calc_saturation_profile
but I had to move all the computations to cell-centred before doing a final interpolation of tau_sat from cell to face. This would inevitably change the results from the reference. Furthermore, there is now an error with calc_propagate_forcing./delete-remote.sh But whatever. I will be squishing this commit message. Committing to not lose my changes.